Add PK variant of Q for testing arbitrary primary keys
This commit extends the support introduced in "Add support for
querying models by primary-key value" to non-integer primary keys by
adding the new syntax PK(value) to denote an equality test for _value_
against an arbitrary model's primary key.  The following three
statements are now equivalent:

   blog = Blog.get(id=1)  # relies on knowledge that Blog's pk is "id"
   blog = Blog.get(1)     # bare-int notation added in earlier commit
   blog = Blog.get(PK(1)) # PK() notation added in this commit

The PK() notation, however, can be used for non-integer primary keys.
The following two statements, then, are now equivalent:

   content = Textstore.get(hash="8b7ec236c")  # assumes pk is called "hash"
   content = Textstore.get(PK("8b7ec236c"))

The PK() notation can be used anywhere Q() notation can be used,
including in where() and filter() methods.  (PK is actually a subclass
of Q.)

The PK() notation takes an optional argument "test" that denotes the
type of test to use; it defaults to "eq" (equality test).  For
example, the following queries are exactly equivalent:

   Blog.filter(PK(10, "lt"))   # find blogs w/ id < 10
   Blog.filter(id__lt=10)